+++ back-end/src/main/java/group16/backend/workspace/repository/WorkspaceRepo.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface WorkspaceRepo extends JpaRepository<Workspace, Integer> {
+public interface WorkspaceRepo extends JpaRepository<Workspace, Long> {
 
 }
